added get operations to return the current state of the UB and Constraint objects in HklCalculation object. Modified tests for this, and also moved the UB calculation operation to hklCalc routes as it made most sense. Modified tests to reflect these changes.
