+++ src/diffcalc_API/errors/HklCalculation.py
@@ -1,6 +1,7 @@
-from typing import Tuple
+from typing import Optional, Tuple, Union
 
 import numpy as np
+from diffcalc.hkl.calc import HklCalculation
 
 from diffcalc_API.errorDefinitions import DiffcalcAPIException, ErrorCodes, allResponses
 
@@ -8,6 +9,7 @@ from diffcalc_API.errorDefinitions import DiffcalcAPIException, ErrorCodes, allR
 class codes(ErrorCodes):
     check_valid_miller_indices = 400
     check_valid_scan_bounds = 400
+    calculate_UB_matrix = 400
 
 
 responses = {code: allResponses[code] for code in np.unique(codes().all_codes())}
@@ -30,5 +32,20 @@ def check_valid_scan_bounds(start: float, stop: float, inc: float):
                 f"numpy range cannot be formed from start: {start}"
                 f" to stop: {stop} in increments of: {inc}"
             ),
+        )
+    return
+
+
+def calculate_UB_matrix(
+    hkl: HklCalculation,
+    firstTag: Optional[Union[int, str]],
+    secondTag: Optional[Union[int, str]],
+) -> None:
+    try:
+        hkl.ubcalc.calc_ub(firstTag, secondTag)
+    except Exception as e:
+        raise DiffcalcAPIException(
+            status_code=codes.calculate_UB_matrix,
+            detail=f"Error calculating UB matrix: {e.__str__()}",
         )
     return
